fix: 구조개선
