+++ src/main/java/shop/wannab/userservice/address/controller/UserAddressController.java
@@ -15,9 +15,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import shop.wannab.userservice.address.domain.dto.UserAddressCreateRequest;
-import shop.wannab.userservice.address.domain.dto.UserAddressResponse;
-import shop.wannab.userservice.address.domain.dto.UserAddressUpdateRequest;
+import shop.wannab.userservice.address.domain.dto.request.UserAddressCreateRequest;
+import shop.wannab.userservice.address.domain.dto.response.UserAddressResponse;
+import shop.wannab.userservice.address.domain.dto.request.UserAddressUpdateRequest;
 import shop.wannab.userservice.address.exception.AlreadyExistsUserAddressException;
 import shop.wannab.userservice.address.exception.UserAddressFullException;
 import shop.wannab.userservice.address.exception.UserAddressNotFoundException;
