+++ src/main/java/shop/wannab/userservice/address/service/Impl/UserAddressServiceImpl.java
@@ -6,9 +6,9 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import shop.wannab.userservice.address.domain.dto.UserAddressCreateRequest;
-import shop.wannab.userservice.address.domain.dto.UserAddressResponse;
-import shop.wannab.userservice.address.domain.dto.UserAddressUpdateRequest;
+import shop.wannab.userservice.address.domain.dto.request.UserAddressCreateRequest;
+import shop.wannab.userservice.address.domain.dto.response.UserAddressResponse;
+import shop.wannab.userservice.address.domain.dto.request.UserAddressUpdateRequest;
 import shop.wannab.userservice.address.domain.entity.UserAddress;
 import shop.wannab.userservice.address.exception.AlreadyExistsUserAddressException;
 import shop.wannab.userservice.address.exception.UserAddressFullException;
