+++ src/main/java/shop/wannab/userservice/auth/controller/AuthController.java
@@ -10,16 +10,16 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import shop.wannab.userservice.auth.controller.request.PaycoLoginRequest;
-import shop.wannab.userservice.auth.controller.request.ReissueRequest;
-import shop.wannab.userservice.auth.controller.request.TokenPayloadRequest;
-import shop.wannab.userservice.auth.controller.request.TokenRequest;
-import shop.wannab.userservice.auth.controller.request.UnlockRequest;
-import shop.wannab.userservice.auth.controller.response.PaycoLoginResponse;
-import shop.wannab.userservice.auth.controller.response.ReissueResponse;
-import shop.wannab.userservice.auth.controller.response.TokenPayloadResponse;
-import shop.wannab.userservice.auth.controller.response.TokenResponse;
-import shop.wannab.userservice.auth.controller.response.UserResponse;
+import shop.wannab.userservice.auth.dto.request.PaycoLoginRequest;
+import shop.wannab.userservice.auth.dto.request.ReissueRequest;
+import shop.wannab.userservice.auth.dto.request.TokenPayloadRequest;
+import shop.wannab.userservice.auth.dto.request.TokenRequest;
+import shop.wannab.userservice.auth.dto.request.UnlockRequest;
+import shop.wannab.userservice.auth.dto.response.PaycoLoginResponse;
+import shop.wannab.userservice.auth.dto.response.ReissueResponse;
+import shop.wannab.userservice.auth.dto.response.TokenPayloadResponse;
+import shop.wannab.userservice.auth.dto.response.TokenResponse;
+import shop.wannab.userservice.auth.dto.response.UserResponse;
 import shop.wannab.userservice.auth.service.AuthService;
 import shop.wannab.userservice.global.Response;
 import shop.wannab.userservice.point.service.PointHistoryService;
