--- src/main/java/shop/wannab/userservice/auth/controller/request/ReissueRequest.java
@@ -1,4 +1,0 @@
-package shop.wannab.userservice.auth.controller.request;
-
-public record ReissueRequest(String refreshToken) {
-}
