--- src/main/java/shop/wannab/userservice/auth/controller/response/ReissueResponse.java
@@ -1,4 +1,0 @@
-package shop.wannab.userservice.auth.controller.response;
-
-public record ReissueResponse(String accessToken) {
-}
