+++ src/main/java/shop/wannab/userservice/auth/dto/request/PaycoLoginRequest.java
@@ -1,4 +1,4 @@
-package shop.wannab.userservice.auth.controller.request;
+package shop.wannab.userservice.auth.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import java.time.LocalDate;
