+++ src/main/java/shop/wannab/userservice/auth/dto/request/ReissueRequest.java
@@ -1,0 +1,4 @@
+package shop.wannab.userservice.auth.dto.request;
+
+public record ReissueRequest(String refreshToken) {
+}
