+++ src/main/java/shop/wannab/userservice/auth/dto/request/UnlockRequest.java
@@ -1,4 +1,4 @@
-package shop.wannab.userservice.auth.controller.request;
+package shop.wannab.userservice.auth.dto.request;
 
 public record UnlockRequest(
         String userId,
