+++ src/main/java/shop/wannab/userservice/auth/dto/response/PaycoLoginResponse.java
@@ -1,4 +1,4 @@
-package shop.wannab.userservice.auth.controller.response;
+package shop.wannab.userservice.auth.dto.response;
 
 public record PaycoLoginResponse(Long id,
                                  String role) {
