+++ src/main/java/shop/wannab/userservice/auth/dto/response/ReissueResponse.java
@@ -1,0 +1,4 @@
+package shop.wannab.userservice.auth.dto.response;
+
+public record ReissueResponse(String accessToken) {
+}
