+++ src/main/java/shop/wannab/userservice/auth/dto/response/TokenPayloadResponse.java
@@ -1,4 +1,4 @@
-package shop.wannab.userservice.auth.controller.response;
+package shop.wannab.userservice.auth.dto.response;
 
 import java.util.Map;
