+++ src/main/java/shop/wannab/userservice/auth/service/AuthService.java
@@ -11,14 +11,14 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 import shop.wannab.userservice.auth.DoorayMessageClient;
-import shop.wannab.userservice.auth.controller.request.PaycoLoginRequest;
-import shop.wannab.userservice.auth.controller.request.SendMessageRequest;
-import shop.wannab.userservice.auth.controller.request.TokenPayloadRequest;
-import shop.wannab.userservice.auth.controller.request.TokenRequest;
-import shop.wannab.userservice.auth.controller.request.UnlockRequest;
-import shop.wannab.userservice.auth.controller.response.PaycoLoginResponse;
-import shop.wannab.userservice.auth.controller.response.TokenPayloadResponse;
-import shop.wannab.userservice.auth.controller.response.TokenResponse;
+import shop.wannab.userservice.auth.dto.request.PaycoLoginRequest;
+import shop.wannab.userservice.auth.dto.request.SendMessageRequest;
+import shop.wannab.userservice.auth.dto.request.TokenPayloadRequest;
+import shop.wannab.userservice.auth.dto.request.TokenRequest;
+import shop.wannab.userservice.auth.dto.request.UnlockRequest;
+import shop.wannab.userservice.auth.dto.response.PaycoLoginResponse;
+import shop.wannab.userservice.auth.dto.response.TokenPayloadResponse;
+import shop.wannab.userservice.auth.dto.response.TokenResponse;
 import shop.wannab.userservice.global.Response;
 import shop.wannab.userservice.user.domain.entity.State;
 import shop.wannab.userservice.user.domain.entity.User;
