+++ src/main/java/shop/wannab/userservice/point/controller/PointController.java
@@ -15,11 +15,11 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import shop.wannab.userservice.point.domain.dto.PageResponse;
-import shop.wannab.userservice.point.domain.dto.PointHistoryResponse;
-import shop.wannab.userservice.point.domain.dto.PointPolicyCreateRequest;
-import shop.wannab.userservice.point.domain.dto.PointPolicyUpdateDTO;
-import shop.wannab.userservice.point.domain.dto.PointUpdateDTO;
+import shop.wannab.userservice.point.domain.dto.response.PageResponse;
+import shop.wannab.userservice.point.domain.dto.response.PointHistoryResponse;
+import shop.wannab.userservice.point.domain.dto.request.PointPolicyCreateRequest;
+import shop.wannab.userservice.point.domain.dto.request.PointPolicyUpdateDTO;
+import shop.wannab.userservice.point.domain.dto.request.PointUpdateDTO;
 import shop.wannab.userservice.point.domain.entity.PointPolicy;
 import shop.wannab.userservice.point.service.PointHistoryService;
 import shop.wannab.userservice.point.service.PointPolicyService;
