--- src/main/java/shop/wannab/userservice/point/domain/dto/PointUpdateDTO.java
@@ -1,4 +1,0 @@
-package shop.wannab.userservice.point.domain.dto;
-
-public record PointUpdateDTO(int amount) {
-}
