+++ src/main/java/shop/wannab/userservice/point/domain/dto/request/PointHistoryRollbackPointDTO.java
@@ -1,4 +1,4 @@
-package shop.wannab.userservice.point.domain.dto;
+package shop.wannab.userservice.point.domain.dto.request;
 
 import lombok.Builder;
 import shop.wannab.userservice.user.domain.entity.User;
