+++ src/main/java/shop/wannab/userservice/point/domain/dto/request/PointUpdateDTO.java
@@ -1,0 +1,4 @@
+package shop.wannab.userservice.point.domain.dto.request;
+
+public record PointUpdateDTO(int amount) {
+}
