+++ src/main/java/shop/wannab/userservice/point/domain/dto/response/PageResponse.java
@@ -1,4 +1,4 @@
-package shop.wannab.userservice.point.domain.dto;
+package shop.wannab.userservice.point.domain.dto.response;
 
 import java.util.List;
 import org.springframework.data.domain.Page;
