+++ src/main/java/shop/wannab/userservice/point/service/PointHistoryServiceImpl.java
@@ -9,10 +9,10 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import shop.wannab.userservice.point.domain.dto.PointHistoryCreateDTO;
-import shop.wannab.userservice.point.domain.dto.PointHistoryResponse;
-import shop.wannab.userservice.point.domain.dto.PointHistoryRollbackPointDTO;
-import shop.wannab.userservice.point.domain.dto.PointUpdateDTO;
+import shop.wannab.userservice.point.domain.dto.request.PointHistoryCreateDTO;
+import shop.wannab.userservice.point.domain.dto.response.PointHistoryResponse;
+import shop.wannab.userservice.point.domain.dto.request.PointHistoryRollbackPointDTO;
+import shop.wannab.userservice.point.domain.dto.request.PointUpdateDTO;
 import shop.wannab.userservice.point.domain.entity.PointHistory;
 import shop.wannab.userservice.point.domain.entity.PointPolicy;
 import shop.wannab.userservice.point.repository.PointHistoryRepository;
