+++ src/main/java/shop/wannab/userservice/point/service/PointPolicyService.java
@@ -1,8 +1,8 @@
 package shop.wannab.userservice.point.service;
 
 import java.util.List;
-import shop.wannab.userservice.point.domain.dto.PointPolicyCreateRequest;
-import shop.wannab.userservice.point.domain.dto.PointPolicyUpdateDTO;
+import shop.wannab.userservice.point.domain.dto.request.PointPolicyCreateRequest;
+import shop.wannab.userservice.point.domain.dto.request.PointPolicyUpdateDTO;
 import shop.wannab.userservice.point.domain.entity.PointPolicy;
 import shop.wannab.userservice.user.domain.entity.Role;
