+++ src/main/java/shop/wannab/userservice/point/service/PointPolicyServiceImpl.java
@@ -5,8 +5,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import shop.wannab.userservice.point.domain.dto.PointPolicyCreateRequest;
-import shop.wannab.userservice.point.domain.dto.PointPolicyUpdateDTO;
+import shop.wannab.userservice.point.domain.dto.request.PointPolicyCreateRequest;
+import shop.wannab.userservice.point.domain.dto.request.PointPolicyUpdateDTO;
 import shop.wannab.userservice.point.domain.entity.PointPolicy;
 import shop.wannab.userservice.point.exception.AccessDeniedException;
 import shop.wannab.userservice.point.exception.PointPolicyAlreadyExistsException;
