+++ src/main/java/shop/wannab/userservice/user/client/CartClient.java
@@ -3,7 +3,7 @@ package shop.wannab.userservice.user.client;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import shop.wannab.userservice.user.domain.dto.CartCreateRequest;
+import shop.wannab.userservice.user.domain.dto.request.CartCreateRequest;
 
 @FeignClient(name = "order-payment-service", contextId = "cartClient")
 public interface CartClient {
