--- src/main/java/shop/wannab/userservice/user/domain/dto/CartCreateRequest.java
@@ -1,5 +1,0 @@
-package shop.wannab.userservice.user.domain.dto;
-
-
-public record CartCreateRequest(Long userId) {
-}
