--- src/main/java/shop/wannab/userservice/user/domain/dto/Status.java
@@ -1,5 +1,0 @@
-package shop.wannab.userservice.user.domain.dto;
-
-public enum Status {
-    SUCCESS, ERROR
-}
