+++ src/main/java/shop/wannab/userservice/user/domain/dto/request/CartCreateRequest.java
@@ -1,0 +1,5 @@
+package shop.wannab.userservice.user.domain.dto.request;
+
+
+public record CartCreateRequest(Long userId) {
+}
