+++ src/main/java/shop/wannab/userservice/user/domain/entity/Status.java
@@ -1,0 +1,5 @@
+package shop.wannab.userservice.user.domain.entity;
+
+public enum Status {
+    SUCCESS, ERROR
+}
