+++ src/main/java/shop/wannab/userservice/user/service/UserServiceImpl.java
@@ -11,12 +11,12 @@ import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.validation.annotation.Validated;
-import shop.wannab.userservice.auth.controller.response.ReissueResponse;
-import shop.wannab.userservice.auth.controller.response.UserResponse;
-import shop.wannab.userservice.point.domain.dto.PointUpdateDTO;
+import shop.wannab.userservice.auth.dto.response.ReissueResponse;
+import shop.wannab.userservice.auth.dto.response.UserResponse;
+import shop.wannab.userservice.point.domain.dto.request.PointUpdateDTO;
 import shop.wannab.userservice.point.exception.FeignClientException;
 import shop.wannab.userservice.user.client.CartClient;
-import shop.wannab.userservice.user.domain.dto.CartCreateRequest;
+import shop.wannab.userservice.user.domain.dto.request.CartCreateRequest;
 import shop.wannab.userservice.user.domain.dto.request.UserCreateRequest;
 import shop.wannab.userservice.user.domain.dto.request.UserUpdateRequest;
 import shop.wannab.userservice.user.domain.dto.response.UserPageResponse;
