+++ src/test/java/shop/wannab/userservice/controller/UserAddressControllerTest.java
@@ -22,9 +22,9 @@ import org.springframework.http.MediaType;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.web.servlet.MockMvc;
 import shop.wannab.userservice.address.controller.UserAddressController;
-import shop.wannab.userservice.address.domain.dto.UserAddressCreateRequest;
-import shop.wannab.userservice.address.domain.dto.UserAddressResponse;
-import shop.wannab.userservice.address.domain.dto.UserAddressUpdateRequest;
+import shop.wannab.userservice.address.domain.dto.request.UserAddressCreateRequest;
+import shop.wannab.userservice.address.domain.dto.response.UserAddressResponse;
+import shop.wannab.userservice.address.domain.dto.request.UserAddressUpdateRequest;
 import shop.wannab.userservice.address.exception.AlreadyExistsUserAddressException;
 import shop.wannab.userservice.address.exception.UserAddressFullException;
 import shop.wannab.userservice.address.exception.UserAddressNotFoundException;
