+++ src/test/java/shop/wannab/userservice/service/AuthServiceTest.java
@@ -24,10 +24,10 @@ import org.springframework.data.redis.core.HashOperations;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.core.ValueOperations;
 import shop.wannab.userservice.auth.DoorayMessageClient;
-import shop.wannab.userservice.auth.controller.request.SendMessageRequest;
-import shop.wannab.userservice.auth.controller.request.TokenRequest;
-import shop.wannab.userservice.auth.controller.request.UnlockRequest;
-import shop.wannab.userservice.auth.controller.response.TokenResponse;
+import shop.wannab.userservice.auth.dto.request.SendMessageRequest;
+import shop.wannab.userservice.auth.dto.request.TokenRequest;
+import shop.wannab.userservice.auth.dto.request.UnlockRequest;
+import shop.wannab.userservice.auth.dto.response.TokenResponse;
 import shop.wannab.userservice.auth.service.AuthService;
 import shop.wannab.userservice.user.domain.entity.State;
 import shop.wannab.userservice.user.domain.entity.User;
