+++ src/test/java/shop/wannab/userservice/service/PointPolicyServiceImplTest.java
@@ -14,8 +14,8 @@ import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
-import shop.wannab.userservice.point.domain.dto.PointPolicyCreateRequest;
-import shop.wannab.userservice.point.domain.dto.PointPolicyUpdateDTO;
+import shop.wannab.userservice.point.domain.dto.request.PointPolicyCreateRequest;
+import shop.wannab.userservice.point.domain.dto.request.PointPolicyUpdateDTO;
 import shop.wannab.userservice.point.domain.entity.PointPolicy;
 import shop.wannab.userservice.point.exception.AccessDeniedException;
 import shop.wannab.userservice.point.exception.PointPolicyAlreadyExistsException;
